+++ client/building_block_definition_version.go
@@ -210,7 +210,7 @@ type MeshBuildingBlockDefinitionVersionSpec struct {
 
 type MeshBuildingBlockDefinitionVersionStatus struct {
 	State                     MeshBuildingBlockDefinitionVersionState `json:"state" tfsdk:"state"`
-	UsageCount                int64                                   `json:"usageCount" tfsdk:"usage_count"`
+	UsageCount                *int64                                  `json:"usageCount,omitempty" tfsdk:"usage_count"`
 	RedactedForNonOwnerAccess bool                                    `json:"redactedForNonOwnerAccess" tfsdk:"-"`
 }